+++ epserde/tests/test_max_size_of.rs
@@ -9,7 +9,7 @@ use epserde::prelude::*;
 #[repr(C)]
 #[repr(align(32))]
 #[repr(align(64))] // The max wins
-#[epserde_zero_copy]
+#[epserde(zero_copy)]
 struct MyStruct64 {
     u: u32,
 }
@@ -17,14 +17,14 @@ struct MyStruct64 {
 #[derive(Epserde, Debug, Clone, Copy, PartialEq, Eq)]
 #[repr(C)]
 #[repr(align(2))]
-#[epserde_zero_copy]
+#[epserde(zero_copy)]
 struct MyStruct2 {
     u: u32,
 }
 
 #[derive(Epserde, Debug, Clone, Copy, PartialEq, Eq)]
 #[repr(C)]
-#[epserde_zero_copy]
+#[epserde(zero_copy)]
 struct MyStruct {
     u: u32,
 }