+++ epserde/tests/test_types.rs
@@ -139,7 +139,7 @@ fn test_vec_vec_array_array_usize() {
 #[test]
 fn test_struct_deep() {
     #[derive(Epserde, Copy, Clone, Debug, PartialEq)]
-    #[epserde_deep_copy]
+    #[epserde(deep_copy)]
     struct Struct {
         a: usize,
         b: usize,
@@ -163,7 +163,7 @@ fn test_struct_deep() {
 fn test_struct_zero() {
     #[derive(Epserde, Copy, Clone, Debug, PartialEq)]
     #[repr(C)]
-    #[epserde_zero_copy]
+    #[epserde(zero_copy)]
     struct Struct {
         a: usize,
         b: usize,
@@ -186,7 +186,7 @@ fn test_struct_zero() {
 #[test]
 fn test_tuple_struct_deep() {
     #[derive(Epserde, Copy, Clone, Debug, PartialEq)]
-    #[epserde_deep_copy]
+    #[epserde(deep_copy)]
     struct Tuple(usize, usize, i32);
     let a = Tuple(0, 1, 2);
     let mut cursor = <AlignedCursor<Aligned16>>::new();
@@ -205,7 +205,7 @@ fn test_tuple_struct_deep() {
 fn test_tuple_struct_zero() {
     #[derive(Epserde, Copy, Clone, Debug, PartialEq)]
     #[repr(C)]
-    #[epserde_zero_copy]
+    #[epserde(zero_copy)]
     struct Tuple(usize, usize, i32);
     let a = Tuple(0, 1, 2);
     let mut cursor = <AlignedCursor<Aligned16>>::new();
@@ -295,7 +295,7 @@ fn test_enum_deep() {
 fn test_enum_zero() {
     #[derive(Epserde, Clone, Copy, Debug, PartialEq)]
     #[repr(C)]
-    #[epserde_zero_copy]
+    #[epserde(zero_copy)]
     enum Data {
         A,
         B(u64),